Added tall_narrow.ppm
This file was causing segmentation fault when coded with reversible option, although it should also cause segmentation fault when irreversible option is used.  Tracking this segmentation fault was hard because, for unknown reasons, it does not always happen. It is a good idea to add it to the list of tests.
The bug was discovered one the master branch, although the development branch is equally affected.  So I am fixing the bug in the development branch after commit 1c56214d71958e9449d0d7535953fb09be535d8f.